requirements.txt
PyQt5==5.12.3
PyQt5-sip==4.19.17
pyserial==3.4
